Fix bugs for deleteCategoryController function

+++ controllers/categoryController.js
@@ -148,19 +148,35 @@ export const singleCategoryController = async (req, res) => {
 };
 
 //delete category
-export const deleteCategoryCOntroller = async (req, res) => {
+export const deleteCategoryController = async (req, res) => {
   try {
     const { id } = req.params;
-    await categoryModel.findByIdAndDelete(id);
+
+    if (!id) {
+      return res.status(422).send({
+        success: false,
+        message: "Category id cannot be empty"
+      });
+    }
+
+    const deletedCategory = await categoryModel.findByIdAndDelete(id);
+
+    if (!deletedCategory) {
+      return res.status(404).send({
+        success: false,
+        message: "Failed to delete because no category is found"
+      });
+    }
+
     res.status(200).send({
       success: true,
-      message: "Categry Deleted Successfully",
+      message: "Category deleted successfully",
     });
   } catch (error) {
     console.log(error);
     res.status(500).send({
       success: false,
-      message: "error while deleting category",
+      message: "Error while deleting category",
       error,
     });
   }
